refactor(mocks): extract secrets+workers modules from mockApi (#991 M3)
Extract the secrets domain (secretsApi: secrets state, requireUnlockedVault/rejectInvisibleLabel/pooledFixtureStatus helpers, listSecrets…deleteAnthropicToken and the vault methods) and the workers domain (workersApi: workers/workerCounter state, workerUpgradeSummary, listWorkers…provisionHostedWorker, and adminListWorkers moved out of the interleaved run/admin block) into their own files.

These two form the ONE accepted import cycle (PRD #991 D4): secrets.ts imports { workers } from "./workers" for deleteAnthropicTokenById (which unbinds workers pinned to a deleted token), and workers.ts imports { secrets } from "./secrets" for setWorkerBindMode (which resolves a token label against the secrets roster). Both cross-reads live inside function bodies; no module-level initializer crosses the modules, so the cycle cannot throw at import (TDZ-safe). secrets is also imported back into index.ts because setJudgeEnabled still reads it there.
