+++ web/src/mocks/mockApi/index.ts
@@ -4,8 +4,6 @@
 // moves, template CRUD, worker tokens, run inputs — all work locally.
 
 import {
-  type AutoStatus,
-  type BindMode,
   type AppSettings,
   type Board,
   type BoardPrefs,
@@ -19,7 +17,6 @@ import {
   type SettingsResponse,
   type SlackLink,
   type UpdateSettingsPayload,
-  type SecretMeta,
   type UserSettings,
   type UserSettingsPatch,
 } from "../../lib/api";
@@ -30,13 +27,9 @@ import { ApiError } from "../../lib/apiError";
 import { isTheme, resolveTheme } from "../../lib/theme";
 import {
   mockAdmin,
-  mockAdminWorkers,
   mockBuildInfo,
   daysAgo,
-  mockMyTokenRateLimits,
   mockRepos,
-  mockSecrets,
-  mockWorkers,
 } from "../data";
 import { nextRunId, patchRun, state } from "../store";
 import { delay, oidcDemo, requireSession, users } from "./shared";
@@ -50,6 +43,8 @@ import { notificationsApi } from "./notifications";
 import { findingsApi } from "./findings";
 import { runsApi } from "./runs";
 import { chatApi } from "./chat";
+import { secretsApi, secrets } from "./secrets";
+import { workersApi } from "./workers";
 
 // ── Settings persistence (demo build) ────────────────────────────────────────
 // The mock persists ONLY the settings maps to localStorage so a hard reload of
@@ -295,20 +290,7 @@ function releaseCheckStatus(): ReleaseCheckStatus {
   };
 }
 
-let secrets: SecretMeta[] = mockSecrets.map((s) => ({ ...s }));
-
-// requireUnlockedVault mirrors the real API: sealing a token needs the vault
-// unlocked (PRD #32), so every create/rotate path throws the same 409 the SPA
-// turns into an unlock prompt.
-function requireUnlockedVault(): void {
-  if (!state.vaultUnlocked) {
-    throw new ApiError(409, "vault is locked; unlock it with your password, then save again", {
-      code: "vault_locked",
-    });
-  }
-}
 let userSettings: UserSettings = loadedSettings.userSettings;
-let workers = mockWorkers.map((w) => ({ ...w }));
 
 // ── Scheduled runs (PRD #241) demo fixtures + helpers ──────────────────────
 // schedulePreviewCap mirrors the server's clamp on the preview N (PRD #241 M4).
@@ -796,8 +778,6 @@ function boardResponse(b: Board): { board: Board } {
     },
   };
 }
-let workerCounter = 0;
-
 
 // sessionBody is the auth/session bootstrap payload: the signed-in user, the
 // current instance labels (PRD #19 M2, PRD #764: the single `uzi` label and the
@@ -829,33 +809,6 @@ function sessionBody() {
           : "opus",
   };
 }
-
-// rejectInvisibleLabel mirrors the server's validateSecretLabel Cf rule (PRD #111).
-//
-// The mock is this repo's BROWSABLE SPEC, and until this existed it accepted labels
-// production rejects — which is how a browser pass managed to store a bidi-override
-// label and demonstrate F12 against a build that was supposed to make it impossible.
-// A mock that disagrees with the API about what is valid teaches the wrong lesson and
-// leaves the new error copy with nowhere to be seen.
-//
-// Control characters are not re-checked here: the real validator rejects them too,
-// but they cannot be typed into the field, so the Cf half is the one a demo exercises.
-function rejectInvisibleLabel(label: string): void {
-  if (/\p{Cf}/u.test(label)) {
-    throw new ApiError(
-      400,
-      "Label must not contain invisible formatting characters (zero-width spaces and joiners, bidirectional overrides, the byte-order mark): they let two different tokens look identical, or make a label read as a different account. This also rules out multi-part emoji such as 👨‍👩‍👧, which are joined by one of these characters, so use a plain name instead",
-    );
-  }
-}
-
-// pooledFixtureStatus is each demo token's eligibility WHEN POOLED, so toggling one
-// off and on again returns it to the state its fixture describes instead of
-// flattening every token to `eligible`.
-const pooledFixtureStatus: Record<string, AutoStatus> = {
-  "sec-never-polled": "no_reading",
-  "sec-low": "below_threshold",
-};
 
 export const mockApi = {
   // ── Auth: instant and fake. Any credentials sign in as the admin. ──────────
@@ -892,16 +845,6 @@ export const mockApi = {
   // the literal "demo" made every badge and the whole Fleet panel unreachable in demo
   // mode. The demo-mode signal does not live here: AppShell renders a separate "demo"
   // pill, so nothing is lost by making this comparable.
-  // PRD #113 M6. Computed LIVE from the demo worker list rather than hardcoded, so the
-  // badge actually clears when a worker is deleted — web-ux needs to see it appear AND
-  // clear, and a constant would only ever show the first half.
-  workerUpgradeSummary: async () => {
-    const attention = workers.filter(
-      (w) => w.upgrade_status === "upgrade_failed" || w.upgrade_status === "outdated",
-    ).length;
-    return delay({ attention, target_release: "0.4.2" });
-  },
-
   // The FULLY-STAMPED fixture (PRD #175), so the demo shows the popover with every
   // row present — a `dev` build here would hide the three fields this PRD exists to
   // add, in the build whose whole job is to show them off.
@@ -1301,201 +1244,7 @@ export const mockApi = {
   ...notificationsApi,
 
   // ── Secrets ─────────────────────────────────────────────────────────────────
-  listSecrets: async () =>
-    delay({
-      // Default first, then by label — the order the server's query returns.
-      secrets: [...secrets]
-        .sort((a, b) =>
-          a.is_default === b.is_default ? a.label.localeCompare(b.label) : a.is_default ? -1 : 1,
-        )
-        .map((s) => ({ ...s })),
-    }),
-  putAnthropicToken: async (_token: string) => {
-    // Mirror the real API: a locked vault cannot seal a new token (PRD #32).
-    requireUnlockedVault();
-    const now = new Date().toISOString();
-    // The D14 alias rotates the DEFAULT, or creates the first one labelled
-    // "default" — exactly what UpsertDefaultUserSecret does server-side.
-    const existing = secrets.find((s) => s.kind === "anthropic_token" && s.is_default);
-    if (existing) {
-      existing.updated_at = now;
-      return delay({ secret: { ...existing } });
-    }
-    const created: SecretMeta = {
-      id: `sec-${Math.random().toString(36).slice(2, 8)}`,
-      kind: "anthropic_token",
-      label: "default",
-      is_default: true,
-      // The user's FIRST/SOLE anthropic_token is born pooled (issue #804) so a
-      // single-token user has a non-empty auto-select pool; token #2+ stays
-      // opt-in. Compute it faithfully off the "no existing anthropic_token" rule,
-      // evaluated BEFORE pushing this row, or the mock teaches the wrong lesson.
-      auto_eligible: secrets.filter((s) => s.kind === "anthropic_token").length === 0,
-      created_at: now,
-      updated_at: now,
-    };
-    secrets.push(created);
-    return delay({ secret: { ...created } });
-  },
-
-  // ── Token CRUD (PRD #104 M2) ───────────────────────────────────────────────
-  createAnthropicToken: async (_token: string, label: string, isDefault: boolean) => {
-    requireUnlockedVault();
-    const trimmed = label.trim();
-    if (trimmed === "") throw new ApiError(400, "label must not be empty");
-    rejectInvisibleLabel(trimmed);
-    const anthropic = () => secrets.filter((s) => s.kind === "anthropic_token");
-    if (anthropic().some((s) => s.label.toLowerCase() === trimmed.toLowerCase())) {
-      throw new ApiError(409, "a token with that label already exists");
-    }
-    // The server FORCES a user's first token to be the default whatever the body
-    // asks (the invisible-token hazard); mirror that here or the mock teaches the
-    // wrong lesson.
-    const first = anthropic().length === 0;
-    const wantDefault = isDefault || first;
-    if (wantDefault) anthropic().forEach((s) => (s.is_default = false));
-    const now = new Date().toISOString();
-    const created: SecretMeta = {
-      id: `sec-${Math.random().toString(36).slice(2, 8)}`,
-      kind: "anthropic_token",
-      label: trimmed,
-      is_default: wantDefault,
-      // The user's FIRST/SOLE anthropic_token is born pooled (issue #804); token
-      // #2+ stays opt-in (auto_eligible false). Mirror the server or the mock
-      // teaches the wrong lesson.
-      auto_eligible: first,
-      created_at: now,
-      updated_at: now,
-    };
-    secrets.push(created);
-    return delay({ secret: { ...created } });
-  },
-  patchAnthropicToken: async (
-    id: string,
-    body: { label?: string; default?: boolean; token?: string },
-  ) => {
-    const row = secrets.find((s) => s.id === id);
-    if (!row) throw new ApiError(404, "token not found");
-    if (body.token !== undefined) requireUnlockedVault();
-    if (body.default === false) {
-      throw new ApiError(400, "cannot clear the default; set another token as default instead");
-    }
-    if (body.label !== undefined) {
-      const trimmed = body.label.trim();
-      if (trimmed === "") throw new ApiError(400, "label must not be empty");
-      rejectInvisibleLabel(trimmed);
-      if (
-        secrets.some(
-          (s) => s.id !== id && s.kind === row.kind && s.label.toLowerCase() === trimmed.toLowerCase(),
-        )
-      ) {
-        throw new ApiError(409, "a token with that label already exists");
-      }
-      row.label = trimmed;
-    }
-    if (body.default === true) {
-      secrets.filter((s) => s.kind === row.kind).forEach((s) => (s.is_default = false));
-      row.is_default = true;
-    }
-    row.updated_at = new Date().toISOString();
-    return delay({ secret: { ...row } });
-  },
-  // The auto-selection pool toggle (PRD #111 M2). It also re-derives the token's
-  // live eligibility, because in the mock that is the only way the chip beside the
-  // toggle can move — and a toggle whose visible consequence never changes is the
-  // silent no-op the real feature exists to make visible.
-  setTokenAutoEligible: async (id: string, autoEligible: boolean) => {
-    const row = secrets.find((s) => s.id === id);
-    if (!row) throw new ApiError(404, "token not found");
-    row.auto_eligible = autoEligible;
-    row.updated_at = new Date().toISOString();
-    const meter = mockMyTokenRateLimits.find((t) => t.secret_id === id);
-    if (meter) {
-      meter.auto_eligible = autoEligible;
-      // Opting OUT is always `not_pooled` — that gate comes first server-side too.
-      // Opting IN restores the token's OWN fixture state rather than hard-coding
-      // `eligible` (web-ux F2): the four states the feature exists for — never
-      // polled, stale, no usage data, low headroom — were unreachable in the demo
-      // because this line asserted every pooled token is pickable, which is the very
-      // thing the chip exists to disprove.
-      //
-      // This does NOT re-implement the gate. The real status is autoselect.Classify's
-      // answer, computed server-side; this restores a fixture value, which is why it
-      // lives here and not in lib/rateLimits.ts.
-      meter.auto_status = autoEligible ? (pooledFixtureStatus[id] ?? "eligible") : "not_pooled";
-    }
-    return delay({ secret: { ...row } });
-  },
-  deleteAnthropicTokenById: async (id: string) => {
-    const row = secrets.find((s) => s.id === id);
-    if (!row) throw new ApiError(404, "token not found");
-    const siblings = secrets.filter((s) => s.kind === row.kind);
-    // D6: the default may not be deleted while others exist — promote first.
-    if (row.is_default && siblings.length > 1) {
-      throw new ApiError(
-        409,
-        "cannot delete the default token while other tokens exist; set another token as default first",
-      );
-    }
-    secrets = secrets.filter((s) => s.id !== id);
-    // The real schema CASCADES: migrations 00078/00079 hang composite FKs off
-    // user_secrets (user_id, id) with ON DELETE SET NULL, so deleting a bound token
-    // unbinds its workers and the judge rather than orphaning them. Without this the
-    // mock left workers reading "spends console-key" forever — and with one token
-    // left the picker is hidden, so there was no way to correct it. Two reasons that
-    // matters beyond tidiness: the shipped Dockerfile.mock demo was showing D5's own
-    // promise being broken, and D5's cascade otherwise has schema-level evidence
-    // only. Mirrored here so a browser can prove the behaviour end to end.
-    workers.forEach((w) => {
-      if (w.anthropic_secret_id === id) {
-        w.anthropic_secret_id = null;
-        w.anthropic_secret_label = null;
-      }
-    });
-    // `state.session` is a COPY, not a reference into `users`, so both have to be
-    // swept or the cascade would be invisible to /me — which is the read every
-    // judge surface actually uses.
-    [...users, state.session].forEach((u) => {
-      if (u && u.judge_anthropic_secret_id === id) {
-        u.judge_anthropic_secret_id = null;
-        u.judge_anthropic_secret_label = null;
-      }
-    });
-    return delay(null);
-  },
-
-  // ── Vault (PRD #32) ───────────────────────────────────────────────────────────
-  // Any non-empty password unlocks in the demo (there is no real crypto); an empty
-  // password is treated as the "wrong password" 403 so the banner's error path is
-  // browsable.
-  vaultUnlock: async (password: string) => {
-    if (password.trim() === "") throw new ApiError(403, "incorrect password");
-    state.vaultUnlocked = true;
-    return delay(null, 150);
-  },
-  // Passphrase-create (PRD #45): min length 12, then the demo vault is unlocked.
-  vaultCreatePassphrase: async (passphrase: string) => {
-    if (passphrase.length < 12) throw new ApiError(400, "passphrase must be at least 12 characters");
-    state.vaultUnlocked = true;
-    return delay(null, 150);
-  },
-  vaultLock: async () => {
-    state.vaultUnlocked = false;
-    return delay(null, 100);
-  },
-  vaultStatus: async () => delay({ unlocked: state.vaultUnlocked }, 40),
-  deleteAnthropicToken: async () => {
-    // D14: the kind-path alias 409s for a multi-token user — they delete by id.
-    const anthropic = secrets.filter((s) => s.kind === "anthropic_token");
-    if (anthropic.length > 1) {
-      throw new ApiError(
-        409,
-        "you have multiple tokens; delete a specific one by id (DELETE /api/me/secrets/anthropic_token/{id})",
-      );
-    }
-    secrets = secrets.filter((s) => s.kind !== "anthropic_token");
-    return delay(null);
-  },
+  ...secretsApi,
   getMySettings: async () => delay({ settings: { ...userSettings } }),
   putMySettings: async (patch: UserSettingsPatch) => {
     // PATCH-like: apply only the fields present in the body, mirroring the real
@@ -1749,158 +1498,13 @@ export const mockApi = {
   },
 
   // ── Workers ─────────────────────────────────────────────────────────────────
-  listWorkers: async () => delay({ workers: workers.map((w) => ({ ...w })) }),
-  createWorker: async (name: string, template?: string) => {
-    const w = {
-      id: `w-new-${++workerCounter}`,
-      name,
-      status: "offline",
-      busy: false,
-      // Hand-run: the user starts this container themselves, so it has no size
-      // (PRD #58). provisionHostedWorker below is the other kind.
-      kind: "external" as const,
-      hosted_size: null,
-      // No runs and no advertised cap until the worker registers (PRD #42).
-      active_runs: 0,
-      max_concurrent_runs: null,
-      // Declared at issuance; reported stays null until the worker registers.
-      template_declared: template ?? null,
-      template_reported: null,
-      version: null,
-      // No version reported until the worker registers, so nothing to compare
-      // against the control-plane release (PRD #113).
-      upgrade_status: "unknown" as const,
-      upgrade_detail: null,
-      upgrade_target: "" as const,
-      upgrade_blocking_container: null,
-      upgrade_blocking_reason: null,
-      upgrade_last_exit_code: null,
-      last_heartbeat_at: null,
-      created_at: new Date().toISOString(),
-      // No resource sample until the worker heartbeats (PRD #49) → no gauges yet.
-      stats_cpu_pct: null,
-      stats_mem_bytes: null,
-      stats_mem_limit_bytes: null,
-      stats_source: null,
-      anthropic_secret_id: null,
-      anthropic_secret_label: null,
-      anthropic_bind_mode: "default" as const,
-      draining_since: null,
-    };
-    workers.push(w);
-    const token = `uzi_wk_${Array.from(crypto.getRandomValues(new Uint8Array(18)), (b) => b.toString(16).padStart(2, "0")).join("")}`;
-    return delay({ worker: { ...w }, token });
-  },
-  deleteWorker: async (id: string) => {
-    workers = workers.filter((w) => w.id !== id);
-    return delay(null);
-  },
-  // PRD #104 M3: rebind a worker to a named token, or clear it with null. Mirrors
-  // the real route's label→id resolution and its 400 for an unknown label, so the
-  // picker's error path is browsable.
-  setWorkerBindMode: async (id: string, mode: BindMode, label: string | null) => {
-    const w = workers.find((x) => x.id === id);
-    if (!w) throw new ApiError(404, "worker not found");
-    // Mirrors the server's refusal of a contradictory pair, so the picker's error
-    // path is browsable in the mock rather than only in production.
-    if (mode !== "pinned" && label !== null && label.trim() !== "") {
-      throw new ApiError(400, "anthropic_token must be null when anthropic_bind_mode is default or auto");
-    }
-    if (mode !== "pinned") {
-      w.anthropic_bind_mode = mode;
-      w.anthropic_secret_id = null;
-      w.anthropic_secret_label = null;
-      return delay({ worker: { ...w } });
-    }
-    if (label === null || label.trim() === "") {
-      throw new ApiError(400, "anthropic_bind_mode=pinned requires a token label in anthropic_token");
-    }
-    const secret = secrets.find(
-      (x) => x.kind === "anthropic_token" && x.label.toLowerCase() === label.trim().toLowerCase(),
-    );
-    if (!secret) throw new ApiError(400, "no Anthropic token with that label");
-    w.anthropic_bind_mode = "pinned";
-    w.anthropic_secret_id = secret.id;
-    w.anthropic_secret_label = secret.label;
-    return delay({ worker: { ...w } });
-  },
-
-  // Hosted workers (PRD #58). The demo is the only place M5 can be seen working: on a
-  // real stack WORKER_HOSTING_ENABLED is off by default, and turning it on gets you a
-  // worker that sits offline forever, because the controller that would run its pod is
-  // M3's. So hosting is hardcoded ON here — a demo of a feature that renders nothing
-  // is not a demo — and quota 2 against one seeded hosted worker puts the whole
-  // journey three clicks away: provision → 2 of 2 → the button disables → delete →
-  // it enables again.
-  // Quota 5 against FOUR seeded hosted workers (PRD #496 added two cordoned demo
-  // workers, w-cordon-eu and w-cordon-idle, on top of the two PRD #113 M5 seeded). The
-  // load-bearing property is unchanged and is why the numbers moved together: there is
-  // exactly ONE slot of headroom, so web-ux can still drive provision -> at quota ->
-  // button disables -> delete -> it enables again, which is the only way to prove the
-  // client-side gate RELEASES rather than merely starting disabled.
-  //
-  // The second seeded worker is the failed roller, which the demo previously could not
-  // show at all — so a browser pass could only ever validate the healthy path.
-  hostedConfig: async () => delay({ enabled: true, quota: 5, ephemeral_enabled: true }),
-  provisionHostedWorker: async (template: string, size: string, docker = false, name?: string) => {
-    const w = {
-      id: `w-hosted-${++workerCounter}`,
-      // Empty name → the server derives one from template + size (handler's
-      // derivedHostedWorkerName), now AWS-style `base.l-<4-hex>`: dot notation,
-      // lowercase t-shirt letter, random hex suffix. The M5 form sends none, so this
-      // is the live path. The real suffix is crypto/rand; the mock stands in a
-      // counter-derived 4-hex from workerCounter (already ++'d for the id) to stay
-      // deterministic-enough for tests — a plausible suffix, not byte-for-byte parity.
-      name:
-        name?.trim() ||
-        `${template}.${size.toLowerCase()}-${(workerCounter & 0xffff).toString(16).padStart(4, "0")}`,
-      // Offline until the controller starts the pod and it registers — the same
-      // lifecycle a hand-run worker has, just with the controller doing the running.
-      status: "offline",
-      busy: false,
-      active_runs: 0,
-      max_concurrent_runs: null,
-      kind: "hosted" as const,
-      hosted_size: size,
-      docker,
-      template_declared: template,
-      template_reported: null,
-      version: null,
-      // No version reported until the worker registers, so nothing to compare
-      // against the control-plane release (PRD #113).
-      upgrade_status: "unknown" as const,
-      upgrade_detail: null,
-      upgrade_target: "" as const,
-      upgrade_blocking_container: null,
-      upgrade_blocking_reason: null,
-      upgrade_last_exit_code: null,
-      last_heartbeat_at: null,
-      created_at: new Date().toISOString(),
-      stats_cpu_pct: null,
-      stats_mem_bytes: null,
-      stats_mem_limit_bytes: null,
-      stats_source: null,
-      anthropic_secret_id: null,
-      anthropic_secret_label: null,
-      anthropic_bind_mode: "default" as const,
-      draining_since: null,
-    };
-    workers.push(w);
-    // { worker } and NOTHING ELSE. Do not mint a token here the way createWorker does
-    // above: the real endpoint's transaction returns none, its response cannot carry
-    // one, and a mock that invents one documents a contract the server structurally
-    // cannot honor. TypeScript will not catch it (delay() infers its own T, so an
-    // extra field type-checks clean) — mockApi.hosted.test.ts is what does.
-    return delay({ worker: { ...w } });
-  },
+  ...workersApi,
 
   ...runsApi,
 
   ...judgeApi,
 
   ...findingsApi,
-
-  adminListWorkers: async () => delay({ workers: mockAdminWorkers.map((w) => ({ ...w })) }),
 
   ...chatApi,
